feat(module): firenvim > nvim in webbrowser
firenvim is a web extension that embedds nvim in the browser.
if you install the extension then you can use doom nvim in the
browser. you specify the font size by setting the guifont size.
for me 25 is a good value but I didn't want to modify the default
value of 15. I haven't used any other gui's for nvim so I don't
know whether it makes sense to change the default to 25 or not.