feat(module): WIP doom fire config. not working
there are some kind of chenanigans with putting the vimL
config in the lua module. maybe it is not that hard to convert
it to lua but as of now there is some problem. At one moment it
worked but then it stopped working and i am not sure why.
BUT when it worked it was AMAZING.