PackerCompile simple
When plugin/packer_compiled.lua is removed, this patch accommodates an
alternative package reconfiguration environment.

After the package updates, Doom-nvim needs to be shut-down manually by
`:q!` and Doom-nvim needs to be started normally.

Since the package reconfiguration under this alternativeackage
environment happens in a minimal less complicated system, there should
be less error messages.

Current packer interface seems to use some kind of locking to wait
before running package updates.  If we find a simpler way to skip this
wait lock, we may be able to further simplify.  Also, if all packages
are in *opt* with proper sequencing, we can make this package update
cycle completely noise free.

Normally, you don't need to use this procedure to update package
selection.  This is here as a backup starting door for debugging
Doom-nvim.

Signed-off-by: Osamu Aoki <osamu@debian.org>