feat(modules): lsp wrapping for new architecture
This adds wrapping of lsp for one language: lua.
It adds two modules, `auto_install` and `lua`. The former toggles auto
installation, and is sort of a marker module. Beyond installing and
configuring nvim-lsp-installer and DAPInstall.nvim, it doesn't really do
anything. It's purpose is to be checked on the lua module which will use
the installers if it is enabled. Otherwise just falls back to $PATH. Since
I'm on NixOS, I can only test the non-auto installation, but it seems to be
downloading properly for the auto version. The lua module also installs
additional packages, namely lua-dev.nvim, which used to be part of the lsp
module. It could add keybinds, but I couldn't think of any. It has a
special autocommand that `dofile`s the config.lua, which actually sets
up lsp (and possibly dap) lazily and only after the first lua file is
opened. This commit is meant as a template/prototype to other language
wrappings.