chore(dockerfile): update archlinux packages
https://hub.docker.com/_/archlinux/?tab=description says:
"Arch Linux is a rolling release distribution, so a full update is recommended
when installing new packages. In other words, we suggest to either execute RUN
pacman -Syu immediately after your FROM statement or as soon as you docker run
into a container."

Without this there are some errors about LIBC_2.36 when installing new
packages brings in packages depending on it, but the installed libc is
still 2.35.

Signed-off-by: Edwin Török <edwin@etorok.net>