feat(largefile): disable slow features on large files
Use LargeFile.vim to disable syntax highlighting on large files.
Gitsigns and treesitter cna also be very slow on large files
(e.g. gigabytes of logfiles), so disable these as well.

By default a file is considered "large" if it either exceeds 10000
lines, or 1MiB. The linecount is not known when first opening the
buffer, hence the need for a file size limit too.

Signed-off-by: Edwin Török <edwin@etorok.net>