Fixed Terminal not opening via hotkey
The key to open the terminal, F4 by default, is not listed in the hotkeys that cause lazy.nvim to load the terminal module.  As a result, the hotkey to open the terminal does not work unless the terminal module has been loaded by lazy.nvim (which, with current default configuration, only happens if the user manually enters :TermExec or :ToggleTerm.